jsonresponse rename to encode
