+++ handlers/handleTaskDone.go
@@ -6,7 +6,7 @@ import (
 	"strconv"
 	"time"
 
-	jsr "github.com/MichaelSBoop/go_final_project/JSONResponse"
+	"github.com/MichaelSBoop/go_final_project/encode"
 	rep "github.com/MichaelSBoop/go_final_project/repeater"
 	"github.com/MichaelSBoop/go_final_project/storage"
 	"github.com/MichaelSBoop/go_final_project/task"
@@ -18,24 +18,24 @@ func HandleTaskDone(s storage.Storage) http.HandlerFunc {
 	return func(w http.ResponseWriter, r *http.Request) {
 		// Проверяем правильность http-метода
 		if r.Method != http.MethodPost {
-			jsr.ErrorJSON(w, fmt.Errorf("incorrect method"), http.StatusBadRequest)
+			encode.ErrorJSON(w, fmt.Errorf("incorrect method"), http.StatusBadRequest)
 			return
 		}
 		var task task.Task
 		// Получаем задачу по id и проверяем на наличие ошибок
 		id := r.URL.Query().Get("id")
 		if id == "" {
-			jsr.ErrorJSON(w, fmt.Errorf("id is required"), http.StatusBadRequest)
+			encode.ErrorJSON(w, fmt.Errorf("id is required"), http.StatusBadRequest)
 			return
 		}
 		_, err := strconv.Atoi(id)
 		if err != nil {
-			jsr.ErrorJSON(w, fmt.Errorf("incorrect id: %v", err), http.StatusBadRequest)
+			encode.ErrorJSON(w, fmt.Errorf("incorrect id: %v", err), http.StatusBadRequest)
 			return
 		}
 		task, err = s.GetTask(id)
 		if err != nil {
-			jsr.ErrorJSON(w, fmt.Errorf("failed to retrieve task: %v", err), http.StatusBadRequest)
+			encode.ErrorJSON(w, fmt.Errorf("failed to retrieve task: %v", err), http.StatusBadRequest)
 			return
 		}
 		if task.Date == "" {
@@ -46,22 +46,22 @@ func HandleTaskDone(s storage.Storage) http.HandlerFunc {
 		if task.Repeat != "" {
 			newDate, err := rep.NextDate(time.Now(), task.Date, task.Repeat)
 			if err != nil {
-				jsr.ErrorJSON(w, err, http.StatusInternalServerError)
+				encode.ErrorJSON(w, err, http.StatusInternalServerError)
 				return
 			}
 			task.Date = newDate
 			if err = s.ChangeTask(task); err != nil {
-				jsr.ErrorJSON(w, fmt.Errorf("failed to change task data: %v", err), http.StatusBadRequest)
+				encode.ErrorJSON(w, fmt.Errorf("failed to change task data: %v", err), http.StatusBadRequest)
 				return
 			}
 		} else {
 			if err := s.DeleteTask(id); err != nil {
-				jsr.ErrorJSON(w, fmt.Errorf("failed to delete task: %v", err), http.StatusBadRequest)
+				encode.ErrorJSON(w, fmt.Errorf("failed to delete task: %v", err), http.StatusBadRequest)
 				return
 			}
 		}
 		// Формулируем JSON для ответа
-		jsonEmpty := jsr.FormulateResponseEmpty()
+		jsonEmpty := encode.FormulateResponseEmpty()
 		w.Header().Set("Content-Type", "application/json; charset=UTF-8")
 		w.WriteHeader(http.StatusOK)
 		_, err = w.Write(jsonEmpty)
