+++ handlers/handleTasks.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 	"net/http"
 
-	jsr "github.com/MichaelSBoop/go_final_project/JSONResponse"
+	"github.com/MichaelSBoop/go_final_project/encode"
 	"github.com/MichaelSBoop/go_final_project/storage"
 )
 
@@ -15,17 +15,17 @@ func HandleTasks(s storage.Storage) http.HandlerFunc {
 	return func(w http.ResponseWriter, r *http.Request) {
 		// Проверяем правильность метода
 		if r.Method != http.MethodGet {
-			jsr.ErrorJSON(w, fmt.Errorf("incorrect method"), http.StatusBadRequest)
+			encode.ErrorJSON(w, fmt.Errorf("incorrect method"), http.StatusBadRequest)
 			return
 		}
 		// Получаем список задач
 		tasks, err := s.GetTasks(limit)
 		if err != nil {
-			jsr.ErrorJSON(w, fmt.Errorf("failed to retrieve tasks from database: %v", err), http.StatusBadRequest)
+			encode.ErrorJSON(w, fmt.Errorf("failed to retrieve tasks from database: %v", err), http.StatusBadRequest)
 			return
 		}
 		// Формируем JSON ответ
-		jsonTasks := jsr.FormulateResponseTasks(tasks)
+		jsonTasks := encode.FormulateResponseTasks(tasks)
 		w.Header().Set("Content-Type", "application/json; charset=UTF-8")
 		w.WriteHeader(http.StatusOK)
 		_, err = w.Write(jsonTasks)
